requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
Django==4.2.7
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
idna==3.10
kombu==5.3.4
Markdown==3.7
MarkupSafe==3.0.2
oauthlib==3.2.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg==3.1.13
pycparser==2.22
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.1.3
